requirements.txt
python-telegram-bot[job-queue]>=21,<23
python-dotenv>=1.0,<2
